增加单位 B/s; 修改部分指标的数据类型

See merge request cloudcare-tools/datakit!323

 [formerly db34d969a85634339c841f4c651ffd5928113c44 [formerly 38953abd938a00257ff9b4d65cf7e37a25cc44f5] [formerly db34d969a85634339c841f4c651ffd5928113c44 [formerly 38953abd938a00257ff9b4d65cf7e37a25cc44f5] [formerly db34d969a85634339c841f4c651ffd5928113c44 [formerly 38953abd938a00257ff9b4d65cf7e37a25cc44f5] [formerly 48e6c3ea06aefc6064037bc2aaff7a425593f7a4 [formerly f9f136591752ea3549f32f954dbd13daff003084]]]]]



Former-commit-id: cf1959175e1b92a2d9fa68fd378cde770957c939 [formerly 22fa9b34b866781ccd87c67695359fb94f770275]
Former-commit-id: da8c7dfa4de9a58a0d4c0937833674787a3ba6c1 [formerly 1fd33d9cf191faa38a05d52d300a9939518e4bd7] [formerly 994965dd0b0c2e4d68a33a954dfc6cdcc0316a28 [formerly 452653376a170471e0f404006b88e7e6460afb50]]
Former-commit-id: 41ff33276cfc4697300258a9e4dfcb99b0dd8575 [formerly a940a7579b58e4baf526c321f113709f51f75864]
Former-commit-id: 0bea63d16b740c41e1f667f562dfe36f187cdcef

+++ plugins/inputs/net/input.go
@@ -196,10 +196,10 @@ func (i *Input) Collect() error {
 		if i.lastStats != nil {
 			if lastIOStat, ok := i.lastStats[name]; ok {
 				if ioStat.BytesSent >= lastIOStat.BytesSent && ts.Unix() > i.lastTime.Unix() {
-					fields["bytes_sent/sec"] = float64(ioStat.BytesSent-lastIOStat.BytesSent) / float64(ts.Unix()-i.lastTime.Unix())
-					fields["bytes_recv/sec"] = float64(ioStat.BytesRecv-lastIOStat.BytesRecv) / float64(ts.Unix()-i.lastTime.Unix())
-					fields["packets_sent/sec"] = float64(ioStat.PacketsSent-lastIOStat.PacketsSent) / float64(ts.Unix()-i.lastTime.Unix())
-					fields["packets_recv/sec"] = float64(ioStat.PacketsRecv-lastIOStat.PacketsRecv) / float64(ts.Unix()-i.lastTime.Unix())
+					fields["bytes_sent/sec"] = int64(ioStat.BytesSent-lastIOStat.BytesSent) / (ts.Unix() - i.lastTime.Unix())
+					fields["bytes_recv/sec"] = int64(ioStat.BytesRecv-lastIOStat.BytesRecv) / (ts.Unix() - i.lastTime.Unix())
+					fields["packets_sent/sec"] = int64(ioStat.PacketsSent-lastIOStat.PacketsSent) / (ts.Unix() - i.lastTime.Unix())
+					fields["packets_recv/sec"] = int64(ioStat.PacketsRecv-lastIOStat.PacketsRecv) / (ts.Unix() - i.lastTime.Unix())
 				}
 			}
 		}
@@ -224,7 +224,7 @@ func (i *Input) Collect() error {
 				sname := strings.ToLower(stat)
 				if _, ok := linuxProtoRate[sname]; ok {
 					if v, ok := fields[pname+"_"+sname]; ok && v.(int64) >= value && ts.Unix() > i.lastTime.Unix() {
-						fields[pname+"_"+sname+"/sec"] = float64(v.(int64)-value) / float64(ts.Unix()-i.lastTime.Unix())
+						fields[pname+"_"+sname+"/sec"] = (v.(int64) - value) / (ts.Unix() - i.lastTime.Unix())
 					}
 				}
